refactor(rss): extract two test blocks into sibling files
services/rss/mod.rs was 2234 LoC, of which the bottom 351 lines were
two `#[cfg(test)] mod NAME { ... }` blocks (`parse_release_tests`,
`evaluate_candidate_tests`). Production logic stays untouched —
match_library_title, parse_release, evaluate_candidate, and the rest
of the RSS pipeline live in mod.rs. Tests move to:

  parse_release_tests.rs        80   parse_release / parse_season_number
                                     coverage on the various fan-sub
                                     title shapes.
  evaluate_candidate_tests.rs  265   evaluate_candidate decision-tree
                                     coverage (preferred-group, blocked-
                                     group, batch fall-through, finished-
                                     series gating, etc.).

mod.rs declarations stay as `#[cfg(test)] mod NAME;` so the module
names + test paths the harness reports remain identical.

Verify: 1233 unit + 7 PR0 + 1 e2e tests all green, fmt + clippy
-D warnings clean.
